Adding Gradle variable to hide font customisation option
