+++ app/build.gradle
@@ -76,6 +76,7 @@ android {
             buildConfigField "boolean", "translations", "false"
             buildConfigField "String", "defaultLocale", "\"en\""
             resValue "string", "authority", applicationId+versionName+"_DHIS2_DEFAULT_AUTHORITY"
+            buildConfigField "boolean", "customFontHidden", "false"
         }
         hnqis {
             applicationId "org.eyeseetea.malariacare.hnqis_ng_bb"
@@ -83,6 +84,7 @@ android {
             buildConfigField "boolean", "translations", "true"
             buildConfigField "String", "defaultLocale", "\"en\""
             resValue "string", "authority", applicationId+versionName+"_DHIS2_DEFAULT_AUTHORITY"
+            buildConfigField "boolean", "customFontHidden", "true"
         }
         surveillance {
             applicationId "org.eyeseetea.malariacare.surveillance_bb"
@@ -90,6 +92,7 @@ android {
             buildConfigField "boolean", "translations", "true"
             buildConfigField "String", "defaultLocale", "\"en\""
             resValue "string", "authority", applicationId+versionName+"_DHIS2_DEFAULT_AUTHORITY"
+            buildConfigField "boolean", "customFontHidden", "false"
         }
     }
