--- app/src/eds/java/org/eyeseetea/malariacare/views/strategy/SettingsActivityStrategy.java
@@ -1,14 +1,0 @@
-package org.eyeseetea.malariacare.views.strategy;
-
-import android.preference.PreferenceScreen;
-import android.support.annotation.NonNull;
-
-import org.eyeseetea.malariacare.strategies.ISettingsActivityStrategy;
-
-public class SettingsActivityStrategy implements ISettingsActivityStrategy {
-
-    @Override
-    public void afterSetupPreferencesScreen(@NonNull PreferenceScreen preferenceScreen) {
-
-    }
-}
