--- app/src/main/java/org/eyeseetea/malariacare/strategies/ISettingsActivityStrategy.java
@@ -1,7 +1,0 @@
-package org.eyeseetea.malariacare.strategies;
-
-import android.preference.PreferenceScreen;
-
-public interface ISettingsActivityStrategy {
-    void afterSetupPreferencesScreen(PreferenceScreen preferenceScreen);
-}
